% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clear_cache.R
\name{clear_cache}
\alias{clear_cache}
\title{Clear quincunx cache of memoised functions}
\usage{
clear_cache()
}
\value{
Returns a logical value, indicating whether the resetting of the
  cache was successful (\code{TRUE}) or not  \code{FALSE}.
}
\description{
quincunx uses memoised functions for the REST API calls. Use this function to
reset the cache.
}
\examples{
clear_cache()

}
