% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-cohorts.R
\docType{class}
\name{cohorts-class}
\alias{cohorts-class}
\title{An S4 class to represent a set of cohorts}
\description{
The cohorts object consists of two slots, each a table
(\code{\link[tibble]{tibble}}), that combined form a relational database of a
subset of PGS Catalog cohorts. Each cohort is an observation (row) in
the \code{cohorts} table --- main table.
}
\section{Slots}{

\describe{
\item{\code{cohorts}}{\describe{
\item{TODO}{TODO}
\item{TODO}{TODO}
}}

\item{\code{pgs_ids}}{A \code{\link[tibble]{tibble}} listing TODO. Columns:
\describe{
\item{TODO}{TODO}
\item{TODO}{TODO}
}}
}}

