% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-cohorts.R
\name{cohorts}
\alias{cohorts}
\title{Constructor for the S4 cohorts object.}
\usage{
cohorts(cohorts = s4cohorts_cohorts_tbl(), pgs_ids = s4cohorts_pgs_ids_tbl())
}
\arguments{
\item{cohorts}{A \code{\link{s4cohorts_cohorts_tbl}} tibble.}

\item{pgs_ids}{A \code{\link{s4cohorts_pgs_ids_tbl}} tibble.}
}
\value{
An object of class \linkS4class{cohorts}.
}
\description{
Constructor for the S4 \linkS4class{cohorts} object.
}
\keyword{internal}
