% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse-publications.R
\name{collect_pgs_ids}
\alias{collect_pgs_ids}
\title{Collect development and evaluation pgs_ids}
\usage{
collect_pgs_ids(tbl_json)
}
\description{
Collect development and evaluation pgs_ids.
}
\keyword{internal}
