% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse-scores.R
\name{collect_samples}
\alias{collect_samples}
\title{Collect samples_variants and samples_training}
\usage{
collect_samples(tbl_json)
}
\description{
Collect samples_variants and samples_training and reference them by a unique
id (sample_id).
}
\keyword{internal}
