% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count.R
\name{count}
\alias{count}
\title{Extract the count field from a JSON response}
\usage{
count(json_string)
}
\arguments{
\item{json_string}{a string.}
}
\value{
An integer value.
}
\description{
This function takes a string with a JSON response and returns the value of
the count field. If it fails to match the pattern then it returns
\code{NA_integer_}.
}
\keyword{internal}
