% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enter.R
\name{enter}
\alias{enter}
\title{Enter into a specific object and discard all other JSON data}
\usage{
enter(tbl_json, obj, column_id_name = paste0(obj, "_id"), iterable = TRUE)
}
\arguments{
\item{tbl_json}{a json string or tbl_json object.}

\item{obj}{name of the object to be entered.}

\item{column_id_name}{name of the id column to be created; by default becomes
the name of the object (\code{obj}) appended by \code{'_id'}.}

\item{iterable}{whether the object's value is an array (\code{TRUE}) or not
(\code{FALSE}).}
}
\value{
a \code{\link[tidyjson]{tbl_json}} object.
}
\description{
This function is almost the same as \code{\link[tidyjson]{enter_object}}.
This function has only the added benefit of automatically adding a column to
the data frame based on the name of the object. If the object's value is a
JSON array, then this function automatically iterates by calling
\code{\link[tidyjson]{gather_array}}. Set \code{iterable} to \code{FALSE} if
the object value is not an array. For more information lookup
\code{\link[tidyjson]{enter_object}}.
}
\keyword{internal}
