% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{filter_by_id}
\alias{filter_by_id}
\title{Filter PGS Catalog objects by identifier.}
\usage{
filter_by_id(x, id)
}
\arguments{
\item{x}{An object of class either\linkS4class{scores},
\linkS4class{publications}, \linkS4class{traits},
\linkS4class{performance_metrics}, or \linkS4class{sample_sets}.}

\item{id}{Identifier.}
}
\value{
Returns an object of class either \linkS4class{scores},
  \linkS4class{publications}, \linkS4class{traits},
  \linkS4class{performance_metrics}, or \linkS4class{sample_sets}.
}
\description{
Use \code{filter_by_id} to filter PGS Catalog objects by their respective
identifier (\code{id}).
}
\keyword{internal}
