% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/first_non_na.R
\name{first_non_na}
\alias{first_non_na}
\title{Returns the position of the first non-NA value}
\usage{
first_non_na(x)
}
\arguments{
\item{x}{An atomic vector.}
}
\description{
Returns the position of the first non-NA value
}
\keyword{internal}
