% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ancestry_categories.R
\name{get_ancestry_categories}
\alias{get_ancestry_categories}
\title{Get PGS Catalog Ancestry Categories and Classes}
\usage{
get_ancestry_categories()
}
\value{
A tibble with ancestry categories, classes and associated
  information. See  \code{\link{ancestry_categories}} for details about each
  column.
}
\description{
Retrieves Ancestry Categories and Classes. This function simply returns the
object \code{\link{ancestry_categories}}.
}
