% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_paginated.R
\name{is_paginated}
\alias{is_paginated}
\title{Is the response paginated?}
\usage{
is_paginated(json_string)
}
\arguments{
\item{json_string}{a string.}
}
\value{
A logical value.
}
\description{
Checks if the response is paginated by checking if a count element exists in
the response.
}
\keyword{internal}
