% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_pgs_id.R
\name{is_pgs_id}
\alias{is_pgs_id}
\title{Is a string a PGS Catalog identifier?}
\usage{
is_pgs_id(str, convert_NA_to_FALSE = TRUE)
}
\arguments{
\item{str}{A character vector of strings.}

\item{convert_NA_to_FALSE}{Whether to treat \code{NA} as \code{NA}
(\code{convert_NA_to_FALSE = FALSE}) or whether to return \code{FALSE} when
an \code{NA} is found (\code{convert_NA_to_FALSE = TRUE}).}
}
\value{
A logical vector.
}
\description{
Find which strings are valid PGS Catalog identifiers (returns \code{TRUE}).
Association IDs are tested against the following regular expression:
\code{^PGS\\\\d{6}$}.
}
\keyword{internal}
