% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/warnings.R
\name{msg_400}
\alias{msg_400}
\title{Pretty printing of 404 error}
\usage{
msg_400(response)
}
\arguments{
\item{response}{A \code{\link[httr]{response}} object.}
}
\value{
This function is run for its side effect: printing.
}
\description{
Pretty printing of 404 error.
}
\keyword{internal}
