% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/warnings.R
\name{msg_empty}
\alias{msg_empty}
\title{Pretty printing of empty response}
\usage{
msg_empty(response)
}
\arguments{
\item{response}{A \code{\link[httr]{response}} object.}
}
\value{
This function is run for its side effect: printing.
}
\description{
Pretty printing of empty response
}
\keyword{internal}
