% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/n_pages.R
\name{n_pages}
\alias{n_pages}
\title{Number of pages}
\usage{
n_pages(count, limit = 50L)
}
\arguments{
\item{count}{total number of results.}

\item{limit}{number of results per page.}
}
\value{
The number of pages, an integer value.
}
\description{
Determine the number of pages to be requested from the total number of
results (\code{count}) and the number of results per page (\code{limit}). The
wording used here --- \code{count} and \code{limit} --- is borrowed from the
PGS Catalog REST API documentation.
}
\keyword{internal}
