% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/warnings.R
\name{no_msg}
\alias{no_msg}
\title{Pretty printing of no error message}
\usage{
no_msg(response)
}
\arguments{
\item{response}{A \code{\link[httr]{response}} object.}
}
\value{
This function is run for its side effect: printing.
}
\description{
Pretty printing of no error message
}
\keyword{internal}
