% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nr_to_na.R
\name{nr_to_na}
\alias{nr_to_na}
\title{Convert NR (Not Recorded) to NA (Not Available)}
\usage{
nr_to_na(x)
}
\arguments{
\item{x}{a character vector.}
}
\value{
a character vector whose \code{'NR'} have been replaced with
  \code{NA}.
}
\description{
This function converts the \code{'NR'} string to \code{NA}.
}
\keyword{internal}
