% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/offsets.R
\name{offsets}
\alias{offsets}
\title{Generate offset values}
\usage{
offsets(count, limit)
}
\arguments{
\item{count}{total number of results.}

\item{limit}{number of results per page.}
}
\value{
Offset values, an integer vector.
}
\description{
Generate offset values to be passed to the PGS Catalog REST API endpoints.
The offset parameter, together with the limit parameter, allow to access to a
desired range of results. The offset parameter specifies the starting index
of the range of results desired. It is a zero based index.
}
\keyword{internal}
