% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/open_in_pubmed.R
\name{open_in_pubmed}
\alias{open_in_pubmed}
\title{Browse PubMed from PubMed identifiers.}
\usage{
open_in_pubmed(pubmed_id)
}
\arguments{
\item{pubmed_id}{A PubMed identifier, either a character or an integer
vector.}
}
\value{
Returns \code{TRUE} if successful. Note however that this
function is run for its side effect.
}
\description{
This function launches the web browser and opens a tab for each PubMed
citation.
}
\examples{
\dontshow{if (interactive()) withAutoprint(\{ # examplesIf}
open_in_pubmed(c('26301688', '30595370'))
\dontshow{\}) # examplesIf}
}
