% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-performance_metrics.R
\docType{class}
\name{performance_metrics-class}
\alias{performance_metrics-class}
\title{An S4 class to represent a set of PGS Catalog Performance Metrics}
\description{
The performance_metrics object consists of nine slots, each a table
(\code{\link[tibble]{tibble}}), that combined form a relational database of a
subset of performance metrics. Each performance metric is an observation
(row) in the \code{scores} table --- main table. All tables have the column
\code{ppm_id} as primary key.
}
\section{Slots}{

\describe{
\item{\code{performance_metrics}}{\describe{
\item{TODO}{TODO}
\item{TODO}{TODO}
}}

\item{\code{publications}}{A \code{\link[tibble]{tibble}} listing TODO. Columns:
\describe{
\item{TODO}{TODO}
\item{TODO}{TODO}
}}

\item{\code{sample_sets}}{A \code{\link[tibble]{tibble}} listing TODO. Columns:
\describe{
\item{TODO}{TODO}
\item{TODO}{TODO}
}}

\item{\code{samples}}{A \code{\link[tibble]{tibble}} listing TODO. Columns:
\describe{
\item{TODO}{TODO}
\item{TODO}{TODO}
}}

\item{\code{demographics}}{A \code{\link[tibble]{tibble}} listing TODO. Columns:
\describe{
\item{TODO}{TODO}
\item{TODO}{TODO}
}}

\item{\code{cohorts}}{A \code{\link[tibble]{tibble}} listing TODO. Columns:
\describe{
\item{TODO}{TODO}
\item{TODO}{TODO}
}}

\item{\code{pgs_effect_sizes}}{A \code{\link[tibble]{tibble}} listing TODO. Columns:
\describe{
\item{TODO}{TODO}
\item{TODO}{TODO}
}}

\item{\code{pgs_classification_metrics}}{A \code{\link[tibble]{tibble}} listing TODO. Columns:
\describe{
\item{TODO}{TODO}
\item{TODO}{TODO}
}}

\item{\code{pgs_other_metrics}}{A \code{\link[tibble]{tibble}} listing TODO. Columns:
\describe{
\item{TODO}{TODO}
\item{TODO}{TODO}
}}
}}

