% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-performance_metrics.R
\name{performance_metrics}
\alias{performance_metrics}
\title{Constructor for the S4 performance_metrics object.}
\usage{
performance_metrics(
  performance_metrics = s4ppm_performance_metrics_tbl(),
  publications = s4ppm_publications_tbl(),
  sample_sets = s4ppm_sample_sets_tbl(),
  samples = s4ppm_samples_tbl(),
  demographics = s4ppm_demographics_tbl(),
  cohorts = s4ppm_pgs_cohorts_tbl(),
  pgs_effect_sizes = s4ppm_pgs_effect_sizes_tbl(),
  pgs_classification_metrics = s4ppm_pgs_classification_metrics_tbl(),
  pgs_other_metrics = s4ppm_pgs_other_metrics_tbl()
)
}
\arguments{
\item{performance_metrics}{A \code{\link{s4ppm_performance_metrics_tbl}} tibble.}

\item{publications}{A \code{\link{s4ppm_publications_tbl}} tibble.}

\item{sample_sets}{A \code{\link{s4ppm_sample_sets_tbl}} tibble.}

\item{samples}{A \code{\link{s4ppm_samples_tbl}} tibble.}

\item{demographics}{A \code{\link{s4ppm_demographics_tbl}} tibble.}

\item{cohorts}{A \code{\link{s4ppm_pgs_cohorts_tbl}} tibble.}

\item{pgs_effect_sizes}{A \code{\link{s4ppm_pgs_effect_sizes_tbl}} tibble.}

\item{pgs_classification_metrics}{A \code{\link{s4ppm_pgs_classification_metrics_tbl}} tibble.}

\item{pgs_other_metrics}{A \code{\link{s4ppm_pgs_other_metrics_tbl}} tibble.}
}
\value{
An object of class \linkS4class{performance_metrics}.
}
\description{
Constructor for the S4 \linkS4class{performance_metrics} object.
}
\keyword{internal}
