% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/id_mapping.R
\name{pgs_to_ppm}
\alias{pgs_to_ppm}
\title{Map PGS identifiers to PPM identifiers}
\usage{
pgs_to_ppm(pgs_id, verbose = FALSE, warnings = TRUE, progress_bar = TRUE)
}
\arguments{
\item{verbose}{A \code{logical} indicating whether the function should be
verbose about the different queries or not.}

\item{warnings}{A \code{logical} indicating whether to print warnings, if
any.}

\item{progress_bar}{Whether to show a progress bar as the queries are
performed.}

\item{ppm_id}{A character vector of PPM identifiers,
e.g., "PGS000001".}
}
\value{
A data frame of two columns: \code{pgs_id} and \code{ppm_id}.
}
\description{
Map PGS identifiers to PPM identifiers.
}
\examples{
\dontrun{
pgs_to_ppm('PGS000001')
pgs_to_ppm(c('PGS000017', 'PGS000042'))
}
}
