% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/id_mapping.R
\name{pss_to_pgs}
\alias{pss_to_pgs}
\title{Map PSS identifiers to PGP identifiers}
\usage{
pss_to_pgs(
  pss_id,
  interactive = TRUE,
  verbose = FALSE,
  warnings = TRUE,
  progress_bar = TRUE
)
}
\arguments{
\item{pss_id}{A character vector of PSS identifiers,
e.g., "PSS000001".}

\item{interactive}{A logical. If all scores are requested, whether to ask
interactively if we really want to proceed.}

\item{verbose}{A \code{logical} indicating whether the function should be
verbose about the different queries or not.}

\item{warnings}{A \code{logical} indicating whether to print warnings, if
any.}

\item{progress_bar}{Whether to show a progress bar as the queries are
performed.}
}
\value{
A data frame of two columns: \code{pss_id} and \code{pgs_id}.
}
\description{
Map PSS identifiers to PGP identifiers. This is a slow function because it
starts by downloading first all Performance Metrics, as this is the linkage
between PSS and PGS.
}
\examples{
pss_to_pgs('PSS000001', interactive = FALSE)

pss_to_pgs(c('PSS000017', 'PSS000042'), interactive = FALSE)

}
