% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-publications.R
\docType{class}
\name{publications-class}
\alias{publications-class}
\title{An S4 class to represent a set of PGS Catalog Publications}
\description{
The publications object consists of two slots, each a table
(\code{\link[tibble]{tibble}}), that combined form a relational database of a
subset of PGS Catalog Publications. Each publication is an observation (row) in
the \code{publications} table --- main table. All tables have the column
\code{pgp_id} as primary key.
}
\section{Slots}{

\describe{
\item{\code{publications}}{\describe{
\item{TODO}{TODO}
\item{TODO}{TODO}
}}

\item{\code{pgs_ids}}{A \code{\link[tibble]{tibble}} listing TODO. Columns:
\describe{
\item{TODO}{TODO}
\item{TODO}{TODO}
}}
}}

