% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-publications.R
\name{publications}
\alias{publications}
\title{Constructor for the S4 publications object.}
\usage{
publications(publications = publications_tbl(), pgs_ids = pgs_ids_tbl())
}
\arguments{
\item{publications}{A \code{\link{publications_tbl}} tibble.}

\item{pgs_ids}{A \code{\link{pgs_ids_tbl}} tibble.}
}
\value{
An object of class \linkS4class{publications}.
}
\description{
Constructor for the S4 \linkS4class{publications} object.
}
\keyword{internal}
