% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-releases.R
\docType{class}
\name{releases-class}
\alias{releases-class}
\title{An S4 class to represent a set of PGS Catalog Releases}
\description{
The releases object consists of four slots, each a table
(\code{\link[tibble]{tibble}}), that combined form a relational database of a
subset of PGS Catalog releases. Each score is an observation (row) in
the \code{scores} table --- main table. All tables have the column
\code{date} as primary key.
}
\section{Slots}{

\describe{
\item{\code{releases}}{\describe{
\item{TODO}{TODO}
\item{TODO}{TODO}
}}

\item{\code{pgs_ids}}{A \code{\link[tibble]{tibble}} listing TODO. Columns:
\describe{
\item{TODO}{TODO}
\item{TODO}{TODO}
}}

\item{\code{ppm_ids}}{A \code{\link[tibble]{tibble}} listing TODO. Columns:
\describe{
\item{TODO}{TODO}
\item{TODO}{TODO}
}}

\item{\code{pgp_ids}}{A \code{\link[tibble]{tibble}} listing TODO. Columns:
\describe{
\item{TODO}{TODO}
\item{TODO}{TODO}
}}
}}

