% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-releases.R
\name{releases}
\alias{releases}
\title{Constructor for the S4 releases object.}
\usage{
releases(
  releases = s4releases_releases_tbl(),
  pgs_ids = s4releases_pgs_ids_tbl(),
  ppm_ids = s4releases_ppm_ids_tbl(),
  pgp_ids = s4releases_pgp_ids_tbl()
)
}
\arguments{
\item{releases}{A \code{s4releases_releases_tbl} tibble.}

\item{pgs_ids}{A \code{s4releases_pgs_ids_tbl} tibble.}

\item{ppm_ids}{A \code{s4releases_ppm_ids_tbl} tibble.}

\item{pgp_ids}{A \code{s4releases_pgp_ids_tbl} tibble.}
}
\value{
An object of class \linkS4class{releases}.
}
\description{
Constructor for the S4 \linkS4class{releases} object.
}
\keyword{internal}
