% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/warnings.R
\name{request_warning}
\alias{request_warning}
\title{Pretty printing of http error}
\usage{
request_warning(response, warnings = FALSE)
}
\arguments{
\item{response}{A \code{\link[httr]{response}} object.}
}
\value{
This function is run for its side effect: printing.
}
\description{
Pretty printing of http error
}
\keyword{internal}
