% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-sample_sets.R
\docType{class}
\name{sample_sets-class}
\alias{sample_sets-class}
\title{An S4 class to represent a set of PGS Catalog Sample Sets}
\description{
The sample_sets object consists of four slots, each a table
(\code{\link[tibble]{tibble}}), that combined form a relational database of a
subset of PGS Catalog sample sets. Each sample set is an observation (row) in
the \code{sample_sets} table --- main table. All tables have the column
\code{pss_id} as primary key.
}
\section{Slots}{

\describe{
\item{\code{sample_sets}}{\describe{
\item{TODO}{TODO}
\item{TODO}{TODO}
}}

\item{\code{samples}}{A \code{\link[tibble]{tibble}} listing TODO. Columns:
\describe{
\item{TODO}{TODO}
\item{TODO}{TODO}
}}

\item{\code{demographics}}{A \code{\link[tibble]{tibble}} listing TODO. Columns:
\describe{
\item{TODO}{TODO}
\item{TODO}{TODO}
}}

\item{\code{cohorts}}{A \code{\link[tibble]{tibble}} listing TODO. Columns:
\describe{
\item{TODO}{TODO}
\item{TODO}{TODO}
}}
}}

