% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-sample_sets.R
\name{sample_sets}
\alias{sample_sets}
\title{Constructor for the S4 sample_sets object.}
\usage{
sample_sets(
  sample_sets = s4pss_sample_sets_tbl(),
  samples = s4pss_samples_tbl(),
  demographics = s4pss_demographics_tbl(),
  cohorts = s4pss_pgs_cohorts_tbl()
)
}
\arguments{
\item{sample_sets}{A \code{\link{s4pss_sample_sets_tbl}} tibble.}

\item{samples}{A \code{\link{s4pss_samples_tbl}} tibble.}

\item{demographics}{A \code{\link{s4pss_demographics_tbl}} tibble.}

\item{cohorts}{A \code{\link{s4pss_pgs_cohorts_tbl}} tibble.}
}
\value{
An object of class \linkS4class{sample_sets}.
}
\description{
Constructor for the S4 \linkS4class{sample_sets} object.
}
\keyword{internal}
