% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-scores.R
\name{scores}
\alias{scores}
\title{Constructor for the S4 scores object.}
\usage{
scores(
  scores = s4scores_scores_tbl(),
  publications = s4scores_publications_tbl(),
  samples = s4scores_samples_tbl(),
  demographics = s4scores_demographics_tbl(),
  cohorts = s4scores_cohorts_tbl(),
  traits = s4scores_traits_tbl()
)
}
\arguments{
\item{scores}{A \code{s4scores_scores_tbl} tibble.}

\item{publications}{A \code{s4scores_publications_tbl} tibble.}

\item{samples}{A \code{s4scores_samples_tbl} tibble.}

\item{demographics}{A \code{s4scores_demographics_tbl} tibble.}

\item{cohorts}{A \code{s4scores_cohorts_tbl} tibble.}

\item{traits}{A \code{s4scores_traits_tbl} tibble.}
}
\value{
An object of class \linkS4class{scores}.
}
\description{
Constructor for the S4 \linkS4class{scores} object.
}
\keyword{internal}
