% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stages.R
\docType{data}
\name{stages}
\alias{stages}
\title{Study stages}
\format{
A data frame with 5 stages (rows) and 4 columns:
\describe{
  \item{stage}{Study stage.}
  \item{symbol}{One-letter symbol for the stage, or a comma separated combination thereof.}
  \item{name}{Stage name.}
  \item{definition}{Stage description.}
}
}
\source{
\url{https://www.pgscatalog.org/docs/ancestry}
}
\usage{
stages
}
\description{
A dataset containing the various study stages assigned to samples in the PGS Catalog.
}
\examples{
stages

}
\keyword{datasets}
