% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/id_mapping.R
\name{study_to_pgs}
\alias{study_to_pgs}
\title{Map GWAS study to PGS identifiers}
\usage{
study_to_pgs(study_id, verbose = FALSE, warnings = TRUE, progress_bar = TRUE)
}
\description{
Map GWAS study to PGS identifiers
}
