% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-trait_categories.R
\name{trait_categories}
\alias{trait_categories}
\title{Constructor for the S4 trait_categories object.}
\usage{
trait_categories(
  trait_categories = s4trait_categories_trait_categories_tbl(),
  traits = s4trait_categories_traits_tbl()
)
}
\arguments{
\item{trait_categories}{A \code{s4trait_categories_trait_categories_tbl} tibble.}

\item{traits}{A \code{s4trait_categories_traits_tbl} tibble.}
}
\value{
An object of class \linkS4class{trait_categories}.
}
\description{
Constructor for the S4 \linkS4class{trait_categories} object.
}
\keyword{internal}
