% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-traits.R
\name{traits}
\alias{traits}
\title{Constructor for the S4 traits object.}
\usage{
traits(
  traits = s4traits_traits_tbl(),
  pgs_ids = s4traits_pgs_ids_tbl(),
  child_pgs_ids = s4traits_child_pgs_ids_tbl(),
  trait_categories = s4traits_trait_categories_tbl(),
  trait_synonyms = s4traits_trait_synonyms_tbl(),
  trait_mapped_terms = s4traits_trait_mapped_terms_tbl()
)
}
\arguments{
\item{traits}{A \code{s4traits_traits_tbl} tibble.}

\item{pgs_ids}{A \code{s4traits_pgs_ids_tbl} tibble.}

\item{child_pgs_ids}{A \code{s4traits_child_pgs_ids_tbl} tibble.}

\item{trait_categories}{A \code{s4traits_trait_categories_tbl} tibble.}

\item{trait_synonyms}{A \code{s4traits_trait_synonyms_tbl} tibble.}

\item{trait_mapped_terms}{A \code{s4traits_trait_mapped_terms_tbl} tibble.}
}
\value{
An object of class \linkS4class{traits}.
}
\description{
Constructor for the S4 \linkS4class{traits} object.
}
\keyword{internal}
