% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/request.R
\name{user_agent_id}
\alias{user_agent_id}
\title{User agent identification}
\usage{
user_agent_id()
}
\value{
An S3 \code{request} object as defined by the package \code{httr}.
}
\description{
Generates an S3 \code{request} object as defined by the package \code{httr},
that is used to identify this package as the user agent in requests to the
PGS REST API. The user agent identification string is: \code{"quincunx: R
Client for the PGS REST API"}.
}
\keyword{internal}
