% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{bind}
\alias{bind}
\title{Bind PGS Catalog objects}
\usage{
bind(x, ...)
}
\arguments{
\item{x}{An object of class: \linkS4class{scores},
\linkS4class{publications}, \linkS4class{traits},
\linkS4class{performance_metrics}, or \linkS4class{sample_sets}.}

\item{...}{Objects of the same class as \code{x}.}
}
\value{
An object of the same class as \code{x}.
}
\description{
Binds together PGS Catalog objects of the same class. Note that
\code{bind()} preserves duplicates whereas
\code{\link[quincunx:setop]{union}} does not.
}
\examples{
# TODO

}
