% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_trait_categories.R
\name{get_trait_categories}
\alias{get_trait_categories}
\title{Get PGS Catalog Trait Categories}
\usage{
get_trait_categories(verbose = FALSE, warnings = TRUE, progress_bar = TRUE)
}
\arguments{
\item{verbose}{A \code{logical} indicating whether the function should be
verbose about the different queries or not.}

\item{warnings}{A \code{logical} indicating whether to print warnings, if any.}

\item{progress_bar}{Whether to show a progress bar indicating download
progress from the REST API server.}
}
\value{
A \linkS4class{trait_categories} object.
}
\description{
Retrieves all trait categories via the PGS Catalog REST API.
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
get_trait_categories(progress_bar = FALSE)
\dontshow{\}) # examplesIf}
}
