% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_trait_categories.R
\name{get_trait_categories}
\alias{get_trait_categories}
\title{Get PGS Catalog Trait Categories}
\usage{
get_trait_categories(verbose = FALSE, warnings = TRUE, progress_bar = TRUE)
}
\arguments{
\item{verbose}{A \code{logical} indicating whether the function should be
verbose about the different queries or not.}

\item{warnings}{A \code{logical} indicating whether to print warnings, if any.}

\item{progress_bar}{Whether to show a progress bar indicating download
progress from the REST API server.}
}
\value{
An \linkS4class{trait_categories} object.
}
\description{
Retrieves trait categories via the PGS Catalog REST API. The REST
API is queried multiple times with the criteria passed as arguments (see
below). By default all trait categories that match the criteria supplied in the
arguments are retrieved: this corresponds to the default option
\code{set_operation} set to \code{'union'}. If you rather have only the
associations that match simultaneously all criteria provided, then set
\code{set_operation} to \code{'intersection'}.
}
\details{
Please note that all search criteria are vectorised, thus allowing for batch
mode search.
}
\examples{
# Coming soon...

}
