% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{n}
\alias{n}
\alias{n,scores-method}
\alias{n,publications-method}
\alias{n,traits-method}
\alias{n,performance_metrics-method}
\alias{n,sample_sets-method}
\alias{n,cohorts-method}
\alias{n,trait_categories-method}
\alias{n,releases-method}
\title{Number of PGS Catalog entities}
\usage{
n(x, unique = FALSE)

\S4method{n}{scores}(x, unique = FALSE)

\S4method{n}{publications}(x, unique = FALSE)

\S4method{n}{traits}(x, unique = FALSE)

\S4method{n}{performance_metrics}(x, unique = FALSE)

\S4method{n}{sample_sets}(x, unique = FALSE)

\S4method{n}{cohorts}(x, unique = FALSE)

\S4method{n}{trait_categories}(x, unique = FALSE)

\S4method{n}{releases}(x, unique = FALSE)
}
\arguments{
\item{x}{A \linkS4class{scores}, \linkS4class{publications},
\linkS4class{traits}, \linkS4class{performance_metrics}, or
\linkS4class{sample_sets} object.}

\item{unique}{Whether to count only unique entries (\code{TRUE}) or not
(\code{FALSE}).}
}
\value{
An integer scalar.
}
\description{
This function returns the number of unique entities in a PGS Catalog object.
}
\examples{
# TODO

}
