% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/open_in_dbsnp.R
\name{open_in_dbsnp}
\alias{open_in_dbsnp}
\title{Browse dbSNP from SNP identifiers.}
\usage{
open_in_dbsnp(variant_id)
}
\arguments{
\item{variant_id}{A variant identifier, a character vector.}
}
\value{
Returns \code{TRUE} if successful. Note however that this
function is run for its side effect.
}
\description{
This function launches the web browser at dbSNP and opens a tab for each SNP
identifier.
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
open_in_dbsnp('rs56261590')
\dontshow{\}) # examplesIf}
}
