% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-scores.R
\docType{class}
\name{scores-class}
\alias{scores-class}
\title{An S4 class to represent a set of PGS Catalog Polygenic Scores}
\description{
The scores object consists of eleven slots, each a table
(\code{\link[tibble]{tibble}}), that combined form a relational database of a
subset of PGS Catalog polygenic scores. Each score is an observation (row) in
the \code{scores} table --- main table. All tables have the column
\code{pgs_id} as primary key.
}
\section{Slots}{

\describe{
\item{\code{scores}}{\describe{
\item{TODO}{TODO}
\item{TODO}{TODO}
}}

\item{\code{publications}}{A \code{\link[tibble]{tibble}} listing TODO. Columns:
\describe{
\item{TODO}{TODO}
\item{TODO}{TODO}
}}

\item{\code{samples}}{A \code{\link[tibble]{tibble}} listing TODO. Columns:
\describe{
\item{TODO}{TODO}
\item{TODO}{TODO}
}}

\item{\code{demographics}}{A \code{\link[tibble]{tibble}} listing TODO. Columns:
\describe{
\item{TODO}{TODO}
\item{TODO}{TODO}
}}

\item{\code{cohorts}}{A \code{\link[tibble]{tibble}} listing TODO. Columns:
\describe{
\item{TODO}{TODO}
\item{TODO}{TODO}
}}

\item{\code{traits}}{A \code{\link[tibble]{tibble}} listing TODO. Columns:
\describe{
\item{TODO}{TODO}
\item{TODO}{TODO}
}}
}}

