% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{setop}
\alias{setop}
\alias{union}
\alias{intersect}
\alias{setdiff}
\alias{setequal}
\title{Set operations on PGS GWAS Catalog objects.}
\usage{
union(x, y, ...)

intersect(x, y, ...)

setdiff(x, y, ...)

setequal(x, y, ...)
}
\arguments{
\item{x, y}{Objects of either class \linkS4class{scores},
\linkS4class{publications}, \linkS4class{traits},
\linkS4class{performance_metrics}, or \linkS4class{sample_sets}.}

\item{...}{other arguments passed on to methods.}
}
\value{
An object of the same class as \code{x} and \code{y}, i.e.,
  \linkS4class{scores}, \linkS4class{publications}, \linkS4class{traits},
  \linkS4class{performance_metrics}, or \linkS4class{sample_sets}.
}
\description{
Performs set union, intersection, and (asymmetric!) difference on two objects
of either class \linkS4class{scores}, \linkS4class{publications},
\linkS4class{traits}, \linkS4class{performance_metrics}, or
\linkS4class{sample_sets}. Note that \code{union()} removes duplicated
entities, whereas \code{\link[quincunx]{bind}()} does not.
}
\examples{
#
# union()
#
# TODO

#
# intersect()
#
# TODO

#
# setdiff()
#
# TODO

#
# setequal()
#
# TODO

}
