% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{subset-performance_metrics}
\alias{subset-performance_metrics}
\alias{[,performance_metrics,missing,missing,missing-method}
\alias{[,performance_metrics,numeric,missing,missing-method}
\alias{[,performance_metrics,character,missing,missing-method}
\title{Subset a performance_metrics object}
\usage{
\S4method{[}{performance_metrics,missing,missing,missing}(x, i, j, ..., drop = FALSE)

\S4method{[}{performance_metrics,numeric,missing,missing}(x, i, j, ..., drop = FALSE)

\S4method{[}{performance_metrics,character,missing,missing}(x, i, j, ..., drop = FALSE)
}
\arguments{
\item{x}{A \linkS4class{performance_metrics} object.}

\item{i}{Position of the identifier or the name of the identifier itself.}

\item{j}{Not used.}

\item{...}{Additional arguments not used here.}

\item{drop}{Not used.}
}
\value{
A \linkS4class{performance_metrics} object.
}
\description{
You can subset \linkS4class{performance_metrics} by identifier or by position using the
\code{`[`} operator.
}
\examples{
# TODO

}
