% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{subset-releases}
\alias{subset-releases}
\alias{[,releases,missing,missing,missing-method}
\alias{[,releases,numeric,missing,missing-method}
\alias{[,releases,character,missing,missing-method}
\alias{[,releases,Date,missing,missing-method}
\title{Subset a releases object}
\usage{
\S4method{[}{releases,missing,missing,missing}(x, i, j, ..., drop = FALSE)

\S4method{[}{releases,numeric,missing,missing}(x, i, j, ..., drop = FALSE)

\S4method{[}{releases,character,missing,missing}(x, i, j, ..., drop = FALSE)

\S4method{[}{releases,Date,missing,missing}(x, i, j, ..., drop = FALSE)
}
\arguments{
\item{x}{A \linkS4class{releases} object.}

\item{i}{Position of the identifier or the name of the identifier itself.}

\item{j}{Not used.}

\item{...}{Additional arguments not used here.}

\item{drop}{Not used.}
}
\value{
A \linkS4class{releases} object.
}
\description{
You can subset \linkS4class{releases} by identifier or by position using the
\code{`[`} operator.
}
\examples{
# Coming soon...

}
\keyword{internal}
