% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-trait_categories.R
\docType{class}
\name{trait_categories-class}
\alias{trait_categories-class}
\title{An S4 class to represent a set of PGS Catalog Trait Categories}
\description{
The trait_categories object consists of two tables (slots) that combined form
a relational database of a subset of PGS Catalog trait categories. Each score
is an observation (row) in the \code{trait_categories} table (first table).
}
\section{Slots}{

\describe{
\item{\code{trait_categories}}{A table of trait categories. Columns:
\describe{
\item{trait_category}{Trait category name.}
}}

\item{\code{traits}}{A table of associated traits. Columns:
\describe{
\item{trait_category}{Trait category name.}
\item{efo_id}{An \href{https://www.ebi.ac.uk/efo/}{EFO} identifier.}
\item{trait}{Trait name.}
\item{description}{Detailed description of the trait from EFO.}
\item{url}{External link to the EFO entry.}
}}
}}

