% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-trait_categories.R
\docType{class}
\name{trait_categories-class}
\alias{trait_categories-class}
\title{An S4 class to represent a set of PGS Catalog Trait Categories}
\description{
The trait_categories object consists of two slots, each a table
(\code{\link[tibble]{tibble}}), that combined form a relational database of a
subset of PGS Catalog trait categories. Each score is an observation (row) in
the \code{trait_categories} table --- main table.
}
\section{Slots}{

\describe{
\item{\code{trait_categories}}{\describe{
\item{TODO}{TODO}
\item{TODO}{TODO}
}}

\item{\code{traits}}{A \code{\link[tibble]{tibble}} listing TODO. Columns:
\describe{
\item{TODO}{TODO}
\item{TODO}{TODO}
}}
}}

