% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-traits.R
\docType{class}
\name{traits-class}
\alias{traits-class}
\title{An S4 class to represent a set of PGS Catalog Traits}
\description{
The traits object consists of six slots, each a table
(\code{\link[tibble]{tibble}}), that combined form a relational database of a
subset of PGS Catalog traits. Each trait is an observation (row) in
the \code{traits} table --- main table. All tables have the column
\code{efo_id} as primary key.
}
\section{Slots}{

\describe{
\item{\code{traits}}{\describe{
\item{TODO}{TODO}
\item{TODO}{TODO}
}}

\item{\code{pgs_ids}}{A \code{\link[tibble]{tibble}} listing TODO. Columns:
\describe{
\item{TODO}{TODO}
\item{TODO}{TODO}
}}

\item{\code{child_pgs_ids}}{A \code{\link[tibble]{tibble}} listing TODO. Columns:
\describe{
\item{TODO}{TODO}
\item{TODO}{TODO}
}}

\item{\code{trait_categories}}{A \code{\link[tibble]{tibble}} listing TODO. Columns:
\describe{
\item{TODO}{TODO}
\item{TODO}{TODO}
}}

\item{\code{trait_synonyms}}{A \code{\link[tibble]{tibble}} listing TODO. Columns:
\describe{
\item{TODO}{TODO}
\item{TODO}{TODO}
}}

\item{\code{trait_mapped_terms}}{A \code{\link[tibble]{tibble}} listing TODO. Columns:
\describe{
\item{TODO}{TODO}
\item{TODO}{TODO}
}}
}}

